% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyPodcastAPI.R
\name{getAuthenticationToken}
\alias{getAuthenticationToken}
\title{Title}
\usage{
getAuthenticationToken(client_id, client_secret_id)
}
\arguments{
\item{client_id}{}

\item{client_secret_id}{}
}
\value{

}
\description{
Title
}
\examples{

}
