% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyPodcastAPI.R
\name{getEpisodeInformation}
\alias{getEpisodeInformation}
\title{Get Episode Information}
\usage{
getEpisodeInformation(episode_id, market = "US")
}
\arguments{
\item{episode_id}{string: The episode's ID}

\item{market}{string: string: (optional, defaults to US) Returns shows that are available in that market}
}
\value{
a DataFrame that contains the name and a brief discription of the episode.
}
\description{
Get Episode Information
}
\examples{
episode_id('4nRWJ76Tu0ceXJj3uJc4D7')
}
