% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyPodcastAPI.R
\name{getRecentEpisodes}
\alias{getRecentEpisodes}
\title{Title}
\usage{
getRecentEpisodes(
  podcast_id,
  explicit = TRUE,
  limit = 5,
  market = "US",
  duration = NA
)
}
\arguments{
\item{podcast_id}{}

\item{explicit}{}

\item{limit}{}

\item{market}{}

\item{duration}{}
}
\value{

}
\description{
Title
}
\examples{

}
